doc and structure redesign
