+++ README.md
@@ -32,13 +32,27 @@ The goal of this project is to provide such component as a library, to be used i
 
 ## Features
 
-- provides a list of tasks in the system
-- reacts on all task lifecycle events 
-- provides configurable caching of tasks
-- creates of read-optimized projections including task-, process- and business data
+- mirroring tasks: provides a list of tasks in the system including all task attributes provided by Camunda BPM Engine  
+- reacts on all task lifecycle events fired by the engine 
+- high performance queries: creates of read-optimized projections 
+  including task-, process- and business data
+- centralized task list: running several Camunda BPM Engines in several applications 
+  is standard for larger companies. From the user's perspective, it is not feasible 
+  to login to several task lists and check for relevant user tasks. The demand for the 
+  centralized task-list arises and can be addressed by `camunda-bpm-taskpool` 
+  if the tasks from several process engines are collected and transmitted over the network.
+- data enrichment: all use cases in which the data is not stored in the process result 
+  in a cascade of queries executed after the task fetch. The task itself has only the 
+  information of the `executionId`, so you have to query the `RuntimeService` 
+  for the execution, load some variables from it and query external systems for further values. 
+  Another approach is presented in the post from Jan Galinski  
+  (https://blog.holisticon.de/2017/08/prozess-und-business-daten-hand-in-hand-mit-camunda-und-jpa/), 
+  but still results in a query on the task fetch. In contrast to that, 
+  the usage of the `camunda-bpm-taskpool` with a data enrichment plugin mechanism 
+  (allowing to plug-in some data enricher on task creation) would allow for caching the additional 
+  business data along with the task information, instead of querying it during task fetch.  
 
 ## Solution Idea
-
 
 ![Architectural whitebox](docs/architecture-collector.png "Architecture Collector")
 
@@ -64,13 +78,8 @@ This library serves as a foundation of several follow-up projects and tools:
 
 - Workload management: apart from the operative task management, the workload management is addressing issues like dynamic task assignment, optimal task distribution, skill-based task assignment, assignment based on presence etc.  For doing so, a task pool to apply all these rules dynamically is required and the `camunda-bpm-taskpool` component can be used for that.
 
-- Centralized task list: running several Camunda BPM Engines in several applications is standard for larger companies. From the user's perspective, it is not feasible to login to several task lists and check for relevant user tasks. The demand for the centralized task-list arises and can be addressed by `camunda-bpm-taskpool` if the tasks from several process engines are collected and transmitted over the network.
-
-- Data enrichment: all use cases in which the data is not stored in the process result in a cascade of queries executed after the task fetch. The task itself has only the information of the `executionId`, so you have to query the `RuntimeService` for the execution, load some variables from it and query external systems for further values. Another approach is presented in the post from Jan Galinski  (https://blog.holisticon.de/2017/08/prozess-und-business-daten-hand-in-hand-mit-camunda-und-jpa/), but still results in a query on the task fetch. In contrast to that, the usage of the `camunda-bpm-taskpool` with a data enrichment plugin mechanism (allowing to plug-in some data enricher on task creation) would allow for caching the additional business data along with the task information, instead of querying it during task fetch.
-
-
 ## Contribution
 
-If you want to contribute to this project, feel free to do so. Start with [contributing guide](/.github/CONTRIBUTING.md).
+If you want to contribute to this project, feel free to do so. Start with [contributing guide](/src/main/asciidoc/contributing.adoc).
