+++ src/main/asciidoc/contributing.adoc
@@ -1,14 +1,14 @@
-# Contributing Guide
+= Contributing Guide
 
 There are several ways in which you may contribute to this project.
 
-* [File issues](https://github.com/holuda-io/camunda-bpm-taskpool/issues)
-* [Submit a pull requests](#submit-a-pull-request)
+* https://github.com/holuda-io/camunda-bpm-taskpool/issues[File issues]
+* Submit a pull requests
 
-Read more on [how to get the project up and running](#project-setup).
+Read more on how to get the project up and running below.
 
 
-## Submit a Pull Request
+== Submit a Pull Request
 
 If you would like to submit a pull request make sure to 
 
@@ -16,35 +16,27 @@ If you would like to submit a pull request make sure to
 - stick to project coding conventions
 
 
-## Project Setup
+== Project Setup
 
-_Perform the following steps to get a development setup up and running._
+Perform the following steps to get a development setup up and running.
 
 - `git clone https://github.com/holunda-io/camunda-bpm-taskpool.git`
 - `cd camunda-bpm-taskpool`
 - `./mvnw clean install`
 
-## Version control
+== Version control
 
 We are using gitflow in our git SCM. That means that you should start from `develop` branch,
 create a `feature/<name>` out of it and once it is completed create a pull request containing 
 it. Please squash your commits before submitting and use semantic commit messages, if possible. 
  
- 
+== Maven Profiles
 
-
-
-## Release management
-
-See [RELEASE.MD](/docs/RELEASE.MD)
-
-## Profiles
-
-### Camunda Version
+=== Camunda Version
 
 You can choose the used Camunda version by specifying the profile camunda-ee or camunda-ce.
 Specify `-Pcamunda-ee` to switch to Camunda Enterprise edition.
 
-### Skip Frontend
+=== Skip Frontend
 
 If you are interested in backend only, specify the `-DskipFrontend` switch.
