+++ src/main/asciidoc/index.adoc
@@ -1,0 +1,12 @@
+= Camunda BPM Taskpool
+
+== Components
+
+:taskpooldir: ../../../taskpool
+:asciidocdir: src/main/asciidoc
+
+include::{taskpooldir}/taskpool-collector/{asciidocdir}/collector.adoc[leveloffset=2]
+
+include::contributing.adoc[leveloffset=1]
+
+include::release.adoc[leveloffset=1]
