+++ taskpool/taskpool-collector/src/main/kotlin/io/holunda/camunda/taskpool/enricher/ProcessVariablesCorrelator.kt
@@ -36,15 +36,3 @@ class ProcessVariablesCorrelator(vararg correlations: ProcessVariableCorrelation
   }
 }
 
-/**
- * Describes correlation between data entries for a business process definition.
- * @param processDefinitionKey process definition key.
- * @param correlations a map from task definition key to correlation map (variableName, entryType).
- * @param globalCorrelations a global (per-process) correlation map (variableName, entryType).
- */
-class ProcessVariableCorrelation(
-  val processDefinitionKey: ProcessDefinitionKey,
-  val correlations: Map<TaskDefinitionKey, Map<String, EntryType>>,
-  val globalCorrelations: Map<String, EntryType> = emptyMap()
-)
-
