+++ taskpool/taskpool-collector/src/main/kotlin/io/holunda/camunda/taskpool/enricher/ProcessVariablesTaskCommandEnricher.kt
@@ -44,18 +44,3 @@ open class ProcessVariablesTaskCommandEnricher(
   }
 }
 
-fun VariableMap.putAllTyped(source: VariableMap) {
-  source.keys.forEach {
-    this.putValueTyped(it, source.getValueTyped(it))
-  }
-}
-
-inline fun VariableMap.filterKeys(predicate: (String) -> Boolean): VariableMap {
-  val result = Variables.createVariables()
-  for (entry in this) {
-    if (predicate(entry.key)) {
-      result[entry.key] = entry.value
-    }
-  }
-  return result
-}
