+++ taskpool/taskpool-collector/src/main/kotlin/io/holunda/camunda/taskpool/sender/gateway/AxonCommandListGateway.kt
@@ -1,7 +1,6 @@
 package io.holunda.camunda.taskpool.sender.gateway
 
 import io.holunda.camunda.taskpool.TaskCollectorProperties
-import io.holunda.camunda.taskpool.api.task.WithTaskId
 import io.holunda.camunda.taskpool.sender.CommandSender
 import org.axonframework.commandhandling.gateway.CommandGateway
 import org.slf4j.Logger
@@ -9,21 +8,22 @@ import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
 
 /**
- * Sends commands via AXON command gateway, only if the sender property is enabled.
+ * Sends  a list commands via AXON command gateway one-by-one, only if the sender property is enabled.
  */
 @Component
-open class AxonCommandGatewayWrapper(
+open class AxonCommandListGateway(
   private val commandGateway: CommandGateway,
   private val properties: TaskCollectorProperties
-) : CommandGatewayWrapper {
+) : CommandListGateway {
 
   private val logger: Logger = LoggerFactory.getLogger(CommandSender::class.java)
 
   /**
    * Sends data to gateway. Ignores any errors, but logs.
    */
-  override fun sendToGateway(commands: List<WithTaskId>) {
+  override fun sendToGateway(commands: List<Any>) {
     if (!commands.isEmpty()) {
+
       val nextCommand = commands.first()
       val remainingCommands = commands.subList(1, commands.size)
