+++ taskpool/taskpool-collector/src/main/kotlin/io/holunda/camunda/taskpool/sender/gateway/CommandListGateway.kt
@@ -1,0 +1,13 @@
+package io.holunda.camunda.taskpool.sender.gateway
+
+/**
+ * Defines a gateway proxy, for sending commands.
+ */
+interface CommandListGateway {
+
+  /**
+   * Sends a list of commands to gateway.
+   */
+  fun sendToGateway(commands: List<Any>)
+
+}
