+++ taskpool/taskpool-collector/src/test/kotlin/io/holunda/camunda/taskpool/itest/TaskCollectorITest.kt
@@ -2,7 +2,7 @@ package io.holunda.camunda.taskpool.itest
 
 import io.holunda.camunda.taskpool.api.task.*
 import io.holunda.camunda.taskpool.api.task.CamundaTaskEvent.Companion.CREATE
-import io.holunda.camunda.taskpool.sender.gateway.AxonCommandGatewayWrapper
+import io.holunda.camunda.taskpool.sender.gateway.AxonCommandListGateway
 import org.awaitility.Awaitility.await
 import org.awaitility.Awaitility.waitAtMost
 import org.camunda.bpm.engine.RepositoryService
@@ -43,7 +43,7 @@ import java.util.concurrent.TimeUnit
 class TaskCollectorITest {
 
   @MockBean
-  lateinit var commandGateway: AxonCommandGatewayWrapper
+  lateinit var commandGateway: AxonCommandListGateway
 
   @Autowired
   lateinit var repositoryService: RepositoryService
