+++ taskpool/taskpool-collector/src/test/resources/application-collector-itest.yml
@@ -9,6 +9,8 @@ camunda:
     database:
       type: h2
     history-level: full
+    metrics:
+      enabled: false
   taskpool:
     engine-eventing:
       enabled: true
